backend/requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
django==4.1
django-appconf==1.0.5
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap4==23.2
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-jquery==3.1.0
django-redis==5.3.0
django-rest-knox==4.2.0
django-select2==8.1.2
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
mysqlclient==2.2.0
numpy==1.25.2
openai==1.12.0
optional-django==0.3.0
pandas==2.0.3
pipreqs==0.4.13
pycparser==2.21
pydantic==2.3.0
pydantic-core==2.6.3
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarg==0.1.9